Feat: Add dynamic date calculation based on user input in GetPatientsCount Api
Introduce a switch statement to calculate the start date for a time range, enhancing flexibility for different date ranges. The user can now specify "day," "week," "month," or "year" as input, and the corresponding start date will be calculated accordingly. In case of an unrecognized input, the default is set to 7 days.